Use uint64_t instead of int and size_t

+++ Region/BasicStatsCalculator.h
@@ -13,7 +13,7 @@ namespace carta {
 
 template <typename T>
 struct BasicStats {
-    size_t num_pixels;
+    uint64_t num_pixels;
     double sum;
     double mean;
     double stdDev;
@@ -22,7 +22,7 @@ struct BasicStats {
     double rms;
     double sumSq;
 
-    BasicStats<T>(size_t num_pixels, double sum, double mean, double stdDev, T min_val, T max_val, double rms, double sumSq);
+    BasicStats<T>(uint64_t num_pixels, double sum, double mean, double stdDev, T min_val, T max_val, double rms, double sumSq);
     BasicStats<T>();
     void join(BasicStats<T>& other);
 };
@@ -31,16 +31,16 @@ template <typename T>
 class BasicStatsCalculator {
     T _min_val, _max_val;
     double _sum, _sum_squares;
-    size_t _num_pixels;
+    uint64_t _num_pixels;
     const std::vector<T>& _data;
 
 public:
     BasicStatsCalculator(const std::vector<T>& data);
     BasicStatsCalculator(BasicStatsCalculator& mm, tbb::split);
 
-    void operator()(const tbb::blocked_range<size_t>& r);
+    void operator()(const tbb::blocked_range<uint64_t>& r);
     void join(BasicStatsCalculator& other); // NOLINT
-    void reduce(const int start, const int end);
+    void reduce(const uint64_t start, const uint64_t end);
 
     BasicStats<T> GetStats() const;
 };
